Change BlobFileId to uint64

+++ dbms/src/Storages/Page/PageDefines.h
@@ -74,7 +74,7 @@ using PageFileIdAndLevels = std::vector<PageFileIdAndLevel>;
 
 using PageSize = UInt64;
 
-using BlobFileId = UInt32;
+using BlobFileId = UInt64;
 using BlobFileOffset = UInt64;
 static constexpr BlobFileId INVALID_BLOBFILE_ID = 0;
 static constexpr BlobFileOffset INVALID_BLOBFILE_OFFSET = std::numeric_limits<BlobFileOffset>::max();
